feat: implement FormPopUp component and integrate CreateStorageForm in Storages page
