+++ src/components/FormPopUp.tsx
@@ -1,0 +1,43 @@
+import { useState, type PropsWithChildren, type ReactNode } from "react";
+
+type Props = {
+    button: ReactNode;
+}
+
+export default function FormPopUp({ children, button }: PropsWithChildren<Props>) {
+    const [showForm, setShowForm] = useState(false);
+
+    return (
+        <>
+            {showForm && (
+                <div className="fixed top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 z-50 w-11/12 max-w-150 overflow-y-auto">
+                    <div className="bg-base-100 rounded-lg shadow-xl p-6 md:p-8">
+                        
+                        {children}
+
+                        <div className="flex gap-3 mt-1">
+                            <button
+                                type="button"
+                                onClick={() => setShowForm(false)}
+                                className="btn btn-ghost flex-1"
+                            >
+                                Cancel
+                            </button>
+                        </div>
+                    </div>
+                </div>
+            )}
+
+            {showForm && (
+                <div
+                    className="fixed inset-0 backdrop-blur-sm bg-black/50 z-40"
+                    onClick={() => setShowForm(false)}
+                />
+            )}
+
+            <div onClick={() => setShowForm(true)}>
+                {button}
+            </div>
+        </>
+    )
+}
