+++ src/components/storage/SettingsFrom.tsx
@@ -1,0 +1,11 @@
+type Props ={
+
+}
+
+export default function SettingsForm({}: Props) {
+    return (
+        <div>
+            <h1>Settings</h1>
+        </div>
+    )
+}
