+++ src/pages/dashboard/Storages.tsx
@@ -2,22 +2,12 @@ import { useEffect, useState } from "react";
 import StorageCard from "../../components/dashboard/StorageCard";
 import CreateButton from "../../components/dashboard/CreateButton";
 import { useStorages } from "shelflife-react-hooks";
+import FormPopUp from "../../components/FormPopUp";
+import CreateStorageForm from "../../components/storage/NewForm";
 
 export default function Storages() {
-    const { storages, fetchStorages, createStorage } = useStorages();
+    const { storages, fetchStorages } = useStorages();
 
-    const [showForm, setShowForm] = useState(false);
-    const [newStorageName, setNewStorageName] = useState("");
-
-    const handleSubmit = async (e: React.SubmitEvent) => {
-        e.preventDefault();
-        if (!newStorageName) return;
-
-        await createStorage({ name: newStorageName });
-        fetchStorages();
-        setNewStorageName("");
-        setShowForm(false);
-    };
 
     useEffect(() => {
         fetchStorages();
@@ -42,63 +32,9 @@ export default function Storages() {
                 )}
             </div>
 
-            {showForm && (
-                <div className="fixed top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 z-50 w-11/12 max-w-150 overflow-y-auto">
-                    <div className="bg-base-100 rounded-lg shadow-xl p-6 md:p-8">
-                        <div className="flex justify-between items-center mb-6">
-                            <h2 className="text-2xl font-bold">Add New Storage</h2>
-                        </div>
-
-                        <form onSubmit={handleSubmit} className="space-y-4">
-                            <div className="form-control">
-                                <div className="inline-grid *:[grid-area:1/1]">
-                                    <div className={`status ${newStorageName ? 'status-success' : 'status-error'} animate-ping me-2`}></div>
-                                    <div className={`status ${newStorageName ? 'status-success' : 'status-error'} me-2`}></div>
-                                </div>
-                                <label className="label">
-                                    <span className="label-text font-semibold me-2">Storage Name</span>
-                                </label>
-                                <input
-                                    type="text"
-                                    maxLength={40}
-                                    name="name"
-                                    placeholder="e.g., Kitchen Pantry"
-                                    className="input input-bordered"
-                                    value={newStorageName}
-                                    onChange={(e) => setNewStorageName(e.target.value)}
-                                    required
-                                />
-                            </div>
-
-                            <div className="flex gap-3 mt-6">
-                                <button
-                                    type="submit"
-                                    disabled={!newStorageName}
-                                    className={`btn flex-1 ${newStorageName ? 'btn-info' : 'btn-disabled'}`}
-                                >
-                                    Create Storage
-                                </button>
-                                <button
-                                    type="button"
-                                    onClick={() => setShowForm(false)}
-                                    className="btn btn-ghost flex-1"
-                                >
-                                    Cancel
-                                </button>
-                            </div>
-                        </form>
-                    </div>
-                </div>
-            )}
-
-            {showForm && (
-                <div
-                    className="fixed inset-0 backdrop-blur-sm bg-black/50 z-40"
-                    onClick={() => setShowForm(false)}
-                />
-            )}
-
-            <CreateButton onClick={() => setShowForm(true)} />
+            <FormPopUp button={<CreateButton  onClick={() => {}}/>} >
+                <CreateStorageForm />
+            </FormPopUp>
         </>
     )
 }
